factor out supplemental chapter ev
